isolate cart page to disable ssr
